apps/front-api/pyproject.toml
[tool.poetry]
name = "worker-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
azure-storage-queue = "^12.11.0"
azure-identity = "^1.17.1"
azure-data-tables = "^12.5.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
